+++ src/macro_impl.jl
@@ -17,10 +17,9 @@ function tasks_macro(forex)
 
     settings = Settings()
 
-    inits_before, inits_names = _maybe_handle_init_block!(forbody.args)
-    tls_names = isnothing(inits_before) ? [] : map(x -> x.args[1], inits_before)
-    
-    _maybe_handle_set_block!(settings, forbody.args)
+    locals_before, local_inner = _maybe_handle_atlocal_block!(forbody.args)
+    tls_names = isnothing(locals_before) ? [] : map(x -> x.args[1], locals_before)
+    _maybe_handle_atset_block!(settings, forbody.args)
 
     forbody = esc(forbody)
     itrng = esc(itrng)
@@ -44,11 +43,13 @@ function tasks_macro(forex)
             tmapreduce(mapping_function, $(settings.reducer), $(itrng); scheduler = $(settings.scheduler)) 
         end
     elseif settings.collect
+        maybe_warn_useless_init(settings)
         quote
             $make_mapping_function
             tmap(mapping_function, $(itrng); scheduler = $(settings.scheduler))
         end
     else
+        maybe_warn_useless_init(settings)
         quote
             $make_mapping_function
             tforeach(mapping_function, $(itrng); scheduler = $(settings.scheduler))
@@ -60,8 +61,8 @@ function tasks_macro(forex)
     result = :(let
     end)
     push!(result.args[2].args, q)
-    if !isnothing(inits_before)
-        for x in inits_before
+    if !isnothing(locals_before)
+        for x in locals_before
             push!(result.args[1].args, x)
         end
     end
@@ -69,10 +70,16 @@ function tasks_macro(forex)
     result
 end
 
+function maybe_warn_useless_init(settings)
+    !isnothing(settings.init) &&
+        @warn("The @set init = ... settings won't have any effect because no reduction is performed.")
+end
+
 Base.@kwdef mutable struct Settings
     scheduler::Expr = :(DynamicScheduler())
     reducer::Union{Expr, Symbol, Nothing} = nothing
     collect::Bool = false
+    init::Union{Expr, Symbol, Nothing} = nothing
 end
 
 function _sym2scheduler(s)
@@ -87,65 +94,65 @@ function _sym2scheduler(s)
     end
 end
 
-function _maybe_handle_init_block!(args)
-    inits_before = nothing
-    init_inner = nothing
+function _maybe_handle_atlocal_block!(args)
+    locals_before = nothing
+    local_inner = nothing
     tlsidx = findfirst(args) do arg
-        arg isa Expr && arg.head == :macrocall && arg.args[1] == Symbol("@init")
+        arg isa Expr && arg.head == :macrocall && arg.args[1] == Symbol("@local")
     end
     if !isnothing(tlsidx)
-        inits_before, init_inner = _unfold_init_block(args[tlsidx].args[3])
+        locals_before, local_inner = _unfold_atlocal_block(args[tlsidx].args[3])
         deleteat!(args, tlsidx)
     end
-    return inits_before, init_inner
+    return locals_before, local_inner
 end
 
-function _unfold_init_block(ex)
-    inits_before = Expr[]
-    inits_names = Expr[]
+function _unfold_atlocal_block(ex)
+    locals_before = Expr[]
+    locals_names = Expr[]
     if ex.head == :(=)
-        initb, init_name = _init_assign_to_exprs(ex)
-        push!(inits_before, initb)
-        push!(inits_names, init_name)
+        localb, localn = _init_assign_to_exprs(ex)
+        push!(locals_before, initb)
+        push!(locals_names, localn)
     elseif ex.head == :block
         tlsexprs = filter(x -> x isa Expr, ex.args) # skip LineNumberNode
         for x in tlsexprs
-            initb, initn = _init_assign_to_exprs(x)
-            push!(inits_before, initb)
-            push!(inits_names,  initn)
+            localb, localn = _init_assign_to_exprs(x)
+            push!(locals_before, localb)
+            push!(locals_names,  localn)
         end
     else
-        throw(ErrorException("Wrong usage of @init. You must either provide a typed assignment or multiple typed assignments in a `begin ... end` block."))
+        throw(ErrorException("Wrong usage of @local. You must either provide a typed assignment or multiple typed assignments in a `begin ... end` block."))
     end
-    return inits_before, inits_names
+    return locals_before, local_inner
 end
 
-function _init_assign_to_exprs(ex)
+function _atlocal_assign_to_exprs(ex)
     left_ex = ex.args[1]
     if left_ex isa Symbol || left_ex.head != :(::)
-        throw(ErrorException("Wrong usage of @init. Expected typed assignment, e.g. `A::Matrix{Float} = rand(2,2)`."))
+        throw(ErrorException("Wrong usage of @local. Expected typed assignment, e.g. `A::Matrix{Float} = rand(2,2)`."))
     end
     tls_sym = esc(left_ex.args[1])
     tls_type = esc(left_ex.args[2])
     tls_def = esc(ex.args[2])
     @gensym tl_storage
-    init_before = :($(tl_storage) = TaskLocalValue{$tls_type}(() -> $(tls_def)))
-    init_name = :($(tls_sym))
-    return init_before, init_name
+    local_before = :($(tl_storage) = TaskLocalValue{$tls_type}(() -> $(tls_def)))
+    local_name = :($(tls_sym))
+    return local_before, local_name
 end
 
-function _maybe_handle_set_block!(settings, args)
+function _maybe_handle_atset_block!(settings, args)
     idcs = findall(args) do arg
         arg isa Expr && arg.head == :macrocall && arg.args[1] == Symbol("@set")
     end
-    isnothing(idcs) && return # no set block found
+    isnothing(idcs) && return # no @set block found
     for i in idcs
         ex = args[i].args[3]
         if ex.head == :(=)
-            _handle_set_single_assign!(settings, ex)
+            _handle_atset_single_assign!(settings, ex)
         elseif ex.head == :block
             exprs = filter(x -> x isa Expr, ex.args) # skip LineNumberNode
-            _handle_set_single_assign!.(Ref(settings), exprs)
+            _handle_atset_single_assign!.(Ref(settings), exprs)
         else
             throw(ErrorException("Wrong usage of @set. You must either provide an assignment or multiple assignments in a `begin ... end` block."))
         end
@@ -157,7 +164,7 @@ function _maybe_handle_set_block!(settings, args)
     end
 end
 
-function _handle_set_single_assign!(settings, ex)
+function _handle_atset_single_assign!(settings, ex)
     if ex.head != :(=)
         throw(ErrorException("Wrong usage of @set. Expected assignment, e.g. `scheduler = StaticScheduler()`."))
     end