+++ src/macros.jl
@@ -11,7 +11,7 @@ Supports reductions (`@set reducer=<reducer function>`) and collecting the resul
 Under the hood, the `for` loop is translated into corresponding parallel
 [`tforeach`](@ref), [`tmapreduce`](@ref), or [`tmap`](@ref) calls.
 
-See also: [`@set`](@ref), [`@init`](@ref)
+See also: [`@set`](@ref), [`@local`](@ref)
 
 ## Examples
 
@@ -67,46 +67,49 @@ Multiple settings are supported, either as separate `@set` statements or via
 * `scheduler` (e.g. `scheduler=:static`): Can be either a [`Scheduler`](@ref) or a `Symbol` (e.g. `:dynamic` or `:static`)
 * `reducer` (e.g. `reducer=+`): Indicates that a reduction should be performed with the provided binary function. See [`tmapreduce`](@ref) for more information.
 * `collect` (e.g. `collect=true`): Indicates that results should be collected (similar to `map`).
+* `init` (e.g. `init=0.0`): Initial value to be used in a reduction (requires `reducer=...`).
 """
 macro set(args...)
     error("The @set macro may only be used inside of a @tasks block.")
 end
 
-"""
-    @init name::T = value
-
-Can be used inside a `@tasks for ... end` block to specify
-[task-local values](@ref TLS) (TLV) via explicitly typed assignments.
-These values will be allocated and de-referenced once per task
-(rather than once per iteration) and can be re-used between different task-local iterations.
-
-There can only be a single `@init` block in a `@tasks for ... end` block. To specify
-multiple TLVs, use `@init begin ... end`. Compared to regular assignments, there are some
-limitations though, e.g. TLVs can't reference each other.
-
-## Examples
-
-```julia
-using OhMyThreads.Tools: taskid
-@tasks for i in 1:10
-    @set scheduler=DynamicScheduler(; nchunks=2)
-    @init x::Vector{Float64} = zeros(3) # TLV
-
-    x .+= 1
-    println(taskid(), " -> ", x)
-end
-```
-
-```julia
-@tasks for i in 1:10
-    @init begin
-        x::Vector{Int64} = rand(Int, 3)
-        M::Matrix{Float64} = rand(3, 3)
+@eval begin
+    """
+        @local name::T = value
+    
+    Can be used inside a `@tasks for ... end` block to specify
+    [task-local values](@ref TLS) (TLV) via explicitly typed assignments.
+    These values will be allocated once per task
+    (rather than once per iteration) and can be re-used between different task-local iterations.
+    
+    There can only be a single `@local` block in a `@tasks for ... end` block. To specify
+    multiple TLVs, use `@local begin ... end`. Compared to regular assignments, there are some
+    limitations though, e.g. TLVs can't reference each other.
+    
+    ## Examples
+    
+    ```julia
+    using OhMyThreads.Tools: taskid
+    @tasks for i in 1:10
+        @set scheduler=DynamicScheduler(; nchunks=2)
+        @local x::Vector{Float64} = zeros(3) # TLV
+    
+        x .+= 1
+        println(taskid(), " -> ", x)
     end
-    # ...
-end
-```
-"""
-macro init(args...)
-    error("The @init macro may only be used inside of a @tasks block.")
+    ```
+    
+    ```julia
+    @tasks for i in 1:10
+        @local begin
+            x::Vector{Int64} = rand(Int, 3)
+            M::Matrix{Float64} = rand(3, 3)
+        end
+        # ...
+    end
+    ```
+    """
+    macro $(Symbol("local"))(args...)
+        error("The @local macro may only be used inside of a @tasks block.")
+    end
 end