+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 OhMyThreads.jl Changelog
 =========================
 
-Version 0.4.7
+Version 0.5.1
 -------------
 
 - ![feature][badge-feature] Added `OhMyThreads.WithTaskLocals` that represents a closure over `TaskLocalValues`, but can have those values materialized as an optimization (using `OhMyThreads.promise_task_local`)
 - ![Enhancement][badge-enhancement] Made `@tasks` use `OhMyThreads.WithTaskLocals` automatically as an optimization.
+
+Version 0.5.0
+-------------
+
+- ![Feature][badge-feature] `@set init = ...` may now be used to specify an initial value for a reduction (only has an effect in conjuction with `@set reducer=...` and triggers a warning otherwise).
+- ![BREAKING][badge-breaking] Within a `@tasks` block, task-local values must from now on be defined via `@local` instead of `@init` (renamed).
+- ![BREAKING][badge-breaking] The (already deprecated) `SpawnAllScheduler` has been dropped.
 
 Version 0.4.6
 -------------